+++ backend-node/CHANGELOG.md
@@ -14,3 +14,6 @@
 * Open route for compute bills
 * Update cascade rounding algorithm to compute complex cases
 * Add prometheus client for metrics support
+* Add CascadeRounding processor
+* Add ComputeBills processor
+* Add esling support